refactor: use one index.ts for each folder instead of multiple
