--- shared/components/HamburgerCheckbox/index.ts
@@ -1,1 +1,0 @@
-export { HamburgerCheckbox } from "./HamburgerCheckbox";
