--- shared/components/LanguageSelector/index.ts
@@ -1,1 +1,0 @@
-export { LanguageSelector } from "./LanguageSelector";
