+++ shared/components/index.ts
@@ -1,0 +1,2 @@
+export { HamburgerCheckbox } from "./HamburgerCheckbox/HamburgerCheckbox";
+export { LanguageSelector } from "./LanguageSelector/LanguageSelector";
