--- shared/sections/About/index.ts
@@ -1,1 +1,0 @@
-export { About } from "./About";
