+++ shared/sections/index.ts
@@ -1,0 +1,3 @@
+export { About } from "./About/About";
+export { Header } from "./Header/Header";
+export { Home } from "./Home/Home";
